+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-154](https://github.com/itk-dev/economics/pull/154)
+  2294: Added worker name field and added to planning overview.
 * [PR-153](https://github.com/itk-dev/economics/pull/153)
   2265: Changed X column in external exported csv.